충돌해결5

+++ src/test/java/shop/mtcoding/filmtalk/seat/SeatRepositoryTest.java
@@ -29,7 +29,7 @@ public class SeatRepositoryTest {
 
     @Test
     public void 쇼타임ID로영화상영관영화관조회(){
-        int id = 1;
+        Long id = 1;
         Optional<Showtime> showtime = showtimeRepository.mFindById(id);
         Timestamp startedAt = showtime.get().getStartedAt();
         LocalDate date = startedAt.toLocalDateTime().toLocalDate();
@@ -155,13 +155,10 @@ public class SeatRepositoryTest {
         //given
         Long showtimeId = 1L;
 
+
         Integer i = seatRepository.mFindCountOfTotalSeat(showtimeId);
 
         System.out.println(i);
     }
-
-
-
-
 
 }
